+++ threedigrid_builder/grid/obstacles.py
@@ -16,7 +16,7 @@ class Obstacles:
     pass
 
 
-def apply_obstacles(lines, obstacles):
+def apply_obstacles(lines, obstacles, levees):
     """Set obstacles on 2D lines by calculating intersection between
     2D flowline line_coords and obstacle linestring. In case of an
     intersection the flowline becomes and LINE_2D_OBSTACLE.
@@ -26,7 +26,13 @@ def apply_obstacles(lines, obstacles):
     Args:
         lines (Lines)
         obstacles (Obstacles)
+        levees (Levees)
     """
+    if len(obstacles) == 0 and len(levees) == 0:
+        return
+    the_geom = np.concatenate([obstacles.the_geom, levees.the_geom])
+    crest_level = np.concatenate([obstacles.crest_level, levees.crest_level])
+
     is_2d = np.isin(lines.kcu, (LineType.LINE_2D_U, LineType.LINE_2D_V))
     coordinates = lines.line_coords[is_2d]
     if not np.isfinite(coordinates).all():
@@ -35,14 +41,14 @@ def apply_obstacles(lines, obstacles):
         )
     lines_tree = pygeos.STRtree(pygeos.linestrings(coordinates.reshape(-1, 2, 2)))
 
-    inscts = lines_tree.query_bulk(obstacles.the_geom, predicate="intersects")
+    inscts = lines_tree.query_bulk(the_geom, predicate="intersects")
     is_u = np.where(lines.kcu == LineType.LINE_2D_U)[0]
     mask = np.isin(is_u, inscts[1, :])
     lines.kcu[is_u[mask]] = LineType.LINE_2D_OBSTACLE_U
     is_v = np.where(lines.kcu == LineType.LINE_2D_V)[0]
     mask = np.isin(is_v, inscts[1, :])
     lines.kcu[is_v[mask]] = LineType.LINE_2D_OBSTACLE_V
-    for i in range(len(obstacles.id)):
+    for i in range(len(crest_level)):
         indices = inscts[1, np.where(inscts[0, :] == i)]
-        lines.flod[indices] = np.fmax(lines.flod[indices], obstacles.crest_level[i])
+        lines.flod[indices] = np.fmax(lines.flod[indices], crest_level[i])
         lines.flou[indices] = lines.flod[indices]